[localize] Refactor to isolate runtime implementations (#1831)
Before
------

Before this PR, all of the localize runtime logic was contained in the main lit-localize.js module, and it was directly integrated into the msg function in a way that wasn't amenable to tree-shaking.

After
-----

Now, the runtime implementation is factored out, and gets installed only when the end-user actually needs the runtime implementation. If an implementation is not installed, the msg function is a pure identity function.

(Note that for the time being, to avoid a breaking change, the main module still exports the runtime configuration function, but in a future breaking change I'll require importing directly from the new init/ modules. This change at least unlocks tree-shaking and creates the new organizational structure.)

Why?
----
- I'm planning to add an SSR-oriented runtime implementation very shortly, which will make msg able to context-switch when rendering multiple locales simultaneously. Since a number of things are different about this implementation, it makes the most sense to have two separate implementations of the msg function.

- It's completely valid to use the msg function even if you don't actually need to setup the localization runtime. Imagine a re-distributable component library, which uses msg to ensure it could be localized. Users who don't need localization shouldn't need to pay the cost of the full runtime -- all they need is an identity msg function.

- I'm also thinking eventually of implementing a "baked" runtime mode, which will be another kind of transformer output where all IDs are hard-coded, so that hashing isn't required. Again it will be preferable to just have a different msg implementation, and to omit all of the hashing code from the import graph.

- It's more readable.
