--- packages/localize-tools/src/fnv1a64.ts
@@ -1,1 +1,0 @@
-../../localize/src/fnv1a64.ts
