--- packages/localize-tools/src/id-generation.ts
@@ -1,1 +1,0 @@
-../../localize/src/id-generation.ts
