+++ packages/localize/.gitignore
@@ -1,8 +1,7 @@
-/lib/
 /node_modules/
 /tests/
 /.tsbuildinfo
 /lit-localize.*
 /examples/*/node_modules/
-/fnv1a64.*
-/id-generation.*
+/internal/
+/init/
