+++ packages/localize/src/internal/deferred.ts
