+++ packages/localize/src/internal/id-generation.ts
